requirements.txt
# Core ML and Data Science Libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Deep Learning Libraries
torch>=1.9.0
torchvision>=0.10.0
albumentations>=1.4.3

# Google Earth Engine Integration
earthengine-api>=0.1.0
geemap>=0.20.0

# Image Processing
Pillow>=8.3.0
opencv-python>=4.5.0
scikit-image>=0.20.0

# HTTP and Data Handling
requests>=2.25.0

# Logging and Utilities
structlog>=21.1.0

# Scientific Computing
scipy>=1.7.0

# Jupyter Support (optional)
jupyter>=1.0.0
ipykernel>=6.0.0
